Shoppingcart with partialview is now working

+++ InmemDb/Models/Order.cs
@@ -5,7 +5,7 @@ namespace InmemDb.Models
     public class Order
     {
         public int OrderId { get; set; }
-        public int OverallAmount { get; set; }
+        public decimal OverallAmount { get; set; }
         public string UserId { get; set; }
 
         public virtual ICollection<DishOrder> DishOrder { get; set; }
